requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18
charset-normalizer==2.0.12
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
requests==2.27.1
smmap==5.0.0
soupsieve==2.3.2.post1
urllib3==1.26.9
